% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_nf_sheet.R
\name{validate_nf_sheet}
\alias{validate_nf_sheet}
\title{validate_nf_sheet}
\usage{
validate_nf_sheet(nf_sample_sheet, nf_pipeline = c("scri", "nf-core"))
}
\arguments{
\item{nf_sample_sheet}{A path to the nextflow cutandrun pipeline sample sheet.}

\item{nf_pipeline}{character string, 'scri' or 'nf-core', to specify which pipeline you are running.}
}
\description{
Validate the nextflow sample sheet for cutandrun pipeline to obtain
mandatory columns: sample_id, sample, single_end, target_or_control,
read1, and read2. In addition, validate the "sample" group to include at
least one target sample
}
\examples{

\dontrun{
validate('nf-core-sample-sheet.csv', 'nf-core')
}

}
\author{
Chao-Jen
}
