% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sample_sheet.R
\name{tidy_sample_sheet}
\alias{tidy_sample_sheet}
\title{tiudy_sample_sheet}
\usage{
tidy_sample_sheet(
  master_sheet,
  sheet = 1,
  unique_id = NULL,
  sample_id = NULL,
  antibody,
  cell_line,
  single_end = FALSE,
  sample = NULL,
  treatment = NULL
)
}
\arguments{
\item{master_sheet}{A path to the master sample sheet}

\item{sheet}{Sheet to read}

\item{unique_id}{Either \code{NULL} or a character string giving a column name of the \emph{unique_id} information}

\item{sample_id}{Either \code{NULL} or a character string giving a column name of the \emph{sample_id} information.}

\item{antibody}{A character string giving a column name in the master sheet that contains the \emph{antibody} information.}

\item{cell_line}{A character string giving a column name in the master sheet that contains of the \emph{sample_id} information.}

\item{single_end}{FALSE if the FASTQ is paired-end or TRUE is the FASTQ is single-ended.}

\item{sample}{A character string giving a column name in the master sheet that contains of the \emph{sample} (Nextflow-specific) information.}

\item{treatment}{A character string giving a column name in the master sheet that contains of the \emph{target_or_control} (Nextflow-specific) information.}
}
\value{
A data frame formatted to Nextflow-specific sample sheet with mandated columns including \code{sample_id}, \code{sample}, \code{singe_end}, \code{target_or_control}, \code{read1}, \code{read2} and addition columns from the master sample sheet.
}
\description{
Process a sample sheet, tidying and converting it
to the Nexflow sample sheet format for CUT&RUN/CUT&Tag-specific
pipeline.
}
\examples{
\dontrun{
master_sheet <- file.path('/Users/cwo11/Projects/tidynoddy/inst/extdata',
                          'JFSe8_Free_CnR_Template.xlsx')
tidy_sample_sheet(master_sheet,
                  sheet = 1,
                  unique_id = 'samp_name',
                  single_end = FALSE,
                  cell_line = 'Cell line',
                  antibody = 'Condition',
                  treatment = NULL,
                  sample = NULL)
}
}
